% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatedmatchObject.R
\name{CreatedmatchObject}
\alias{CreatedmatchObject}
\title{CreatedmatchObject}
\usage{
CreatedmatchObject(raw.data, batch.id, project = "dmatchProject",
  PCA = PCA)
}
\arguments{
\item{raw.data}{A list of pairwise samples for batch effects correction}

\item{batch.id}{Batch.id which are used to denote those two samples in the previous fastSVD step}

\item{project}{The name of the project}

\item{PCA}{The output from fastSVD}
}
\value{
A dmatch class object which have slots storing raw.data, batch.id, PCA and etc information
}
\description{
Create a dmatch class object
}
\author{
Mengjie Chen
}
