% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_groups.R
\name{cut_groups}
\alias{cut_groups}
\title{cut_groups}
\usage{
cut_groups(object, K, method = "ward.D")
}
\arguments{
\item{object}{A dmatch class object}

\item{K}{Number of cell types to cut}

\item{method}{The agglomeration method to be used for hierarchical cluster analysis}
}
\value{
A dmatch class object which have slots storing raw.data, batch.id, PCA, and more information
}
\description{
Cut cells in the samples into major cell types
}
\author{
Mengjie Chen
}
