% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_back_to_original_space.R
\name{move_back_to_original_space}
\alias{move_back_to_original_space}
\title{move_back_to_original_space}
\usage{
move_back_to_original_space(object)
}
\arguments{
\item{object}{A dmatch class object}
}
\value{
A dmatch class object which have slots storing raw.data, batch.id, PCA, and more information
}
\description{
Move cells in the PC space to the original high-dimensional space
}
\author{
Mengjie Chen
}
