% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAPlot.R
\name{PCAPlot}
\alias{PCAPlot}
\title{PCAPlot}
\usage{
PCAPlot(PCA, PCs.to.plot, filename)
}
\arguments{
\item{PCA}{The list returned by fastSVD}

\item{PCs.to.plot}{Which two PCs to visualize}

\item{filename}{The path and name of the PCA plot returned}
}
\value{
A png file for the visualization of SVD results
}
\description{
Visualize the result of SVD
}
\author{
Mengjie Chen
}
