% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrected.R
\name{plot_corrected}
\alias{plot_corrected}
\title{plot_corrected}
\usage{
plot_corrected(object, filename = NULL, pcs.plot, celltypes.plot)
}
\arguments{
\item{object}{A dmatch class object.}

\item{filename}{The path and name of the png file. Default is NULL and the PC plot is printed directly.}

\item{pcs.plot}{Which two PCs to plot.}

\item{celltypes.plot}{Which cell clusters to plot.}
}
\value{
Return graphs of samples in PC space after batch effects correction.
}
\description{
PC plots for corrected samples.
}
\author{
Mengjie Chen, Qi Zhan
}
