% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_to_reference_panel.R
\name{projection_to_reference_panel}
\alias{projection_to_reference_panel}
\title{projection_to_reference_panel}
\usage{
projection_to_reference_panel(object, Reference, CorMethod = "pearson",
  use.genes.threshold = NULL)
}
\arguments{
\item{object}{A dmatch class object}

\item{Reference}{The reference panel}

\item{CorMethod}{The distance metric used to calculate the distance between cells in the samples and the primary cell lines in the reference panel}

\item{use.genes.threshold}{The threshold for highly variable genes used for calculating the Pearson Correlation between cells in the samples and the primary cell lines in the reference panel}
}
\value{
A dmatch class object which have slots storing raw.data, batch.id, PCA, and more information
}
\description{
Project the pairwise to the reference panel to identify separate clusters in the samples
}
\author{
Mengjie Chen
}
