% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_alignment_by_2D.R
\name{run_alignment_by_2D}
\alias{run_alignment_by_2D}
\title{run_alignment_by_2D}
\usage{
run_alignment_by_2D(object, quantile = 0.95, K = 30, selected = NULL,
  Nclust = NULL, steps = 20, gra_steps = 10, NCell = 100)
}
\arguments{
\item{object}{A dmatch class object}

\item{K}{The number of PCs for correcting batch effects}

\item{NCell}{The smallest number of cells that a selected cluster should have. Default is 100, and recommend no less than 5% of the sample size}
}
\value{
A dmatch class object which have slots storing raw.data, batch.id, PCA, and more information
}
\description{
Correct batch effects
}
\author{
Mengjie Chen
}
