% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_clusters.R
\name{select_clusters}
\alias{select_clusters}
\title{select_clusters}
\usage{
select_clusters(object, quantile = 0.95)
}
\arguments{
\item{object}{A dmatch class object.}

\item{quantile}{The minimum number of the data points regarded as good points.}
}
\value{
A dmatch class object which have slots storing raw.data, batch.id, PCA, and more information. Specifically, select.clusters slot stores information for shapiro test result and numbers of cells in each cluster in each sample.
}
\description{
Apply the shapiro test for normality for each cluster in each sample, as well as summarize the cell numbers for each cluster in each sample. Those two will guide the selection of anchors-some shared cell clusters across the samples, which will help align the to-be-corrected sample to the reference sample.
}
\author{
Mengjie Chen, Qi Zhan
}
