% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatedmatchObject.R
\name{CreatedmatchObject}
\alias{CreatedmatchObject}
\title{CreatedmatchObject}
\usage{
CreatedmatchObject(raw.data, batch.id, project = "dmatchProject",
  PCA = PCA)
}
\arguments{
\item{raw.data}{A list of pairwise samples for batch effects correction, each of which is a raw read count dataset.}

\item{batch.id}{Batch.ids which are used to denote those two samples in the previous fastSVD step.}

\item{project}{Project name (string).}

\item{PCA}{The output from fastSVD, i.e., the list returned by fastSVD function.}
}
\value{
Initializes the dmatch object. Return a dmatch class object which have slots storing raw.data, batch.id, PCA, and etc. This step only includes the cells and genes used in fastSVD step for the samples provided in the raw.data.
}
\description{
Create a dmatch class object
}
\author{
Mengjie Chen, Qi Zhan
}
