% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSVD.R
\name{fastSVD}
\alias{fastSVD}
\title{fastSVD}
\usage{
fastSVD(samples.list, nPC = 30)
}
\arguments{
\item{samples.list}{A list of input datasets preprocessed by Seurat.}

\item{nPC}{Total number of PCs to compute and store (30 by default).}
}
\value{
A list consists of: PCs; Loadings; Centers (centers of samples before SVD); batch.id.forPC (batch id of cells in the samples for SVD); Raw representing the input for SVD; cells are the cell names for cells that are included and used for SVD; genes are the gene names for genes that are included and used for SVD.
}
\description{
Perform SVD. The input datasets are the already preprocessed (filtering) and global-scale-nomalized datasets from Seurat.
}
\author{
Mengjie Chen, Qi Zhan
}
